Singleton and Logger
The singleton was pretty straightforward but the Logger, well i did a
lot of research and took on several implementations, all of them were
overly complicated and created memory leaks (unbelievable). So i'm
running my own super simple logger. It logs on files for now, will log
on console and have levels of logs. It will not be performance oriented
but it will be used to debug mostly