Fonts are now working
There is a fancy font manager that allows to load and set fonts to be
used, and the text renders according to the current font. There is some
cleanup to do. Logs are being logged better now. Will begin to read a
game engine architecture book and refactor things to be better. All this
at a snails pace